Refactor lock crate into SRP modules
