+++ crates/xchecker-lock/src/lib.rs
@@ -10,594 +10,34 @@
 //! - `flow.lock`: reproducibility envelope for the governed flow boundary
 //! - `promotions/<phase>.lock`: trust object for artifacts promoted through a gate
 
+mod paths;
+mod primitives;
+mod state_locks;
+
+#[cfg(any(test, feature = "test-utils"))]
+pub use paths::set_thread_home_for_tests;
+
+pub use primitives::{
+    DriftPair, FlowContext, FlowLock, FlowLockDrift, LockDrift, LockError, LockInfo,
+    PromotionArtifact, PromotionLock, RunContext, XCheckerLock,
+};
+
 use anyhow::Result;
-use camino::Utf8PathBuf;
-use chrono::{DateTime, Utc};
 use fd_lock::RwLock;
-use serde::{Deserialize, Serialize};
-use std::cell::RefCell;
 use std::fs;
 use std::io::{self, Write};
 use std::path::{Path, PathBuf};
 use std::process;
 use std::time::{SystemTime, UNIX_EPOCH};
 
-// Thread-local override used only in tests to avoid process-global env races.
-thread_local! {
-    static THREAD_HOME: RefCell<Option<Utf8PathBuf>> = const { RefCell::new(None) };
-}
+#[cfg(test)]
+use chrono::DateTime;
+#[cfg(test)]
+use paths::with_isolated_home;
+use paths::{ensure_dir_all, spec_root};
 
 /// Default age threshold for considering a lock stale (in seconds)
 const DEFAULT_STALE_THRESHOLD_SECS: u64 = 3600; // 1 hour
-
-/// Lock information stored in the lock file
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct LockInfo {
-    /// Process ID that created the lock
-    pub pid: u32,
-    /// Process start time (seconds since UNIX epoch)
-    pub start_time: u64,
-    /// Timestamp when the lock was created (seconds since UNIX epoch)
-    pub created_at: u64,
-    /// Spec ID being locked
-    pub spec_id: String,
-    /// xchecker version that created the lock
-    pub xchecker_version: String,
-}
-
-/// `XChecker` lockfile for reproducibility tracking (schema v1)
-/// Pins model, CLI version, and schema version to detect drift
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct XCheckerLock {
-    /// Schema version for this lockfile format
-    pub schema_version: String,
-    /// RFC3339 UTC timestamp when the lockfile was created
-    pub created_at: DateTime<Utc>,
-    /// Full model name that was used (e.g., "haiku")
-    pub model_full_name: String,
-    /// Claude CLI version that was used
-    pub claude_cli_version: String,
-}
-
-/// Context for current run to compare against lockfile
-#[derive(Debug, Clone)]
-pub struct RunContext {
-    pub model_full_name: String,
-    pub claude_cli_version: String,
-    pub schema_version: String,
-}
-
-/// Context for a governed flow boundary.
-///
-/// This intentionally pins promoted execution contracts rather than scratch
-/// reasoning. The goal is to make replay / resume deterministic at promotion
-/// points without freezing exploratory thought inside a stage.
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct FlowContext {
-    pub flow_version: String,
-    pub stage_graph_version: String,
-    pub gate_set_version: String,
-    pub prompt_pack_version: String,
-    pub adapter_version: String,
-    pub provider_policy: String,
-    pub tool_context_version: String,
-}
-
-/// Reproducibility lock for the governed flow boundary (`flow.lock`).
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct FlowLock {
-    /// Schema version for this flow lock format
-    pub schema_version: String,
-    /// RFC3339 UTC timestamp when the lock was created
-    pub created_at: DateTime<Utc>,
-    /// Version of the flow pack or flow definition
-    pub flow_version: String,
-    /// Version of the stage graph / phase topology
-    pub stage_graph_version: String,
-    /// Version of the gate set used at promotion points
-    pub gate_set_version: String,
-    /// Version of the prompt pack used by the flow
-    pub prompt_pack_version: String,
-    /// Version of the adapter / harness boundary
-    pub adapter_version: String,
-    /// Capability or provider selection policy applied to the flow
-    pub provider_policy: String,
-    /// Version of tool permissions / MCP context configuration
-    pub tool_context_version: String,
-}
-
-/// Drift information for a `FlowLock`.
-#[derive(Debug, Clone, Serialize, Deserialize, Default, PartialEq, Eq)]
-pub struct FlowLockDrift {
-    pub flow_version: Option<DriftPair>,
-    pub stage_graph_version: Option<DriftPair>,
-    pub gate_set_version: Option<DriftPair>,
-    pub prompt_pack_version: Option<DriftPair>,
-    pub adapter_version: Option<DriftPair>,
-    pub provider_policy: Option<DriftPair>,
-    pub tool_context_version: Option<DriftPair>,
-}
-
-/// Artifact frozen by a promotion lock.
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct PromotionArtifact {
-    /// Artifact path relative to the spec directory
-    pub path: String,
-    /// Canonical BLAKE3 hash of the promoted artifact
-    pub blake3_canonicalized: String,
-}
-
-/// Trust object for artifacts that crossed a gate and became promoted state.
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct PromotionLock {
-    /// Schema version for this promotion lock format
-    pub schema_version: String,
-    /// RFC3339 UTC timestamp when the promotion happened
-    pub promoted_at: DateTime<Utc>,
-    /// Spec identifier this promotion belongs to
-    pub spec_id: String,
-    /// Stage / phase whose artifacts were promoted
-    pub phase: String,
-    /// Winning candidate identifier, if selection happened among alternatives
-    pub selected_candidate_id: Option<String>,
-    /// Gate name that approved the promotion
-    pub gate_name: String,
-    /// Version of the gate logic used to approve the promotion
-    pub gate_version: String,
-    /// Actor or system identity that approved the promotion
-    pub approved_by: Option<String>,
-    /// Receipt path that produced the promotable output
-    pub parent_receipt_path: Option<String>,
-    /// Packet lineage identifiers carried into this promotion
-    pub parent_packet_lineage: Vec<String>,
-    /// Promoted artifacts and their content hashes
-    pub artifacts: Vec<PromotionArtifact>,
-    /// Non-fatal warnings carried with the promotion event
-    pub warnings: Vec<String>,
-}
-
-/// Drift pair showing locked vs current value
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct DriftPair {
-    /// Value from lockfile
-    pub locked: String,
-    /// Current value
-    pub current: String,
-}
-
-/// Lock drift information
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct LockDrift {
-    /// Model full name drift
-    pub model_full_name: Option<DriftPair>,
-    /// Claude CLI version drift
-    pub claude_cli_version: Option<DriftPair>,
-    /// Schema version drift
-    pub schema_version: Option<DriftPair>,
-}
-
-/// Lock errors for file locking operations
-#[derive(Debug, thiserror::Error)]
-pub enum LockError {
-    #[error(
-        "Concurrent execution detected for spec '{spec_id}' (PID {pid}, created {created_ago} ago)"
-    )]
-    ConcurrentExecution {
-        spec_id: String,
-        pid: u32,
-        created_ago: String,
-    },
-
-    #[error(
-        "Stale lock detected for spec '{spec_id}' (PID {pid}, age {age_secs}s). Use --force to override"
-    )]
-    StaleLock {
-        spec_id: String,
-        pid: u32,
-        age_secs: u64,
-    },
-
-    #[error("Lock file is corrupted or invalid: {reason}")]
-    CorruptedLock { reason: String },
-
-    #[error("Failed to acquire lock: {reason}")]
-    AcquisitionFailed { reason: String },
-
-    #[error("Failed to release lock: {reason}")]
-    ReleaseFailed { reason: String },
-
-    #[error("IO error during lock operation: {0}")]
-    Io(#[from] io::Error),
-}
-
-/// Write file atomically using a temporary file and atomic rename
-///
-/// This is a simplified version of atomic_write that doesn't depend on xchecker-utils
-fn write_file_atomic(path: &Utf8PathBuf, content: &str) -> Result<(), io::Error> {
-    let parent = path
-        .parent()
-        .ok_or_else(|| io::Error::new(io::ErrorKind::NotFound, "No parent directory"))?;
-
-    // Ensure parent directory exists
-    fs::create_dir_all(parent)?;
-
-    // Create temporary file in the same directory
-    let temp_path = parent.join(format!(".{}.tmp", path.file_name().unwrap_or("file")));
-
-    // Write content to temporary file
-    fs::write(&temp_path, content)?;
-
-    // Atomically rename temporary file to target path
-    fs::rename(&temp_path, path)?;
-
-    Ok(())
-}
-
-/// Get the spec root directory for a given spec ID
-///
-/// This is a simplified version of paths::spec_root that doesn't depend on xchecker-utils
-fn xchecker_home() -> Utf8PathBuf {
-    if let Some(tl) = THREAD_HOME.with(|tl| tl.borrow().clone()) {
-        return tl;
-    }
-    if let Ok(p) = std::env::var("XCHECKER_HOME") {
-        return Utf8PathBuf::from(p);
-    }
-    Utf8PathBuf::from(".xchecker")
-}
-
-/// Get the spec root directory for a given spec ID
-///
-/// This mirrors xchecker-utils path resolution to keep lock paths consistent.
-fn spec_root(spec_id: &str) -> Utf8PathBuf {
-    xchecker_home().join("specs").join(spec_id)
-}
-
-/// Ensure a directory exists, creating it if necessary
-///
-/// This is a simplified version of paths::ensure_dir_all that doesn't depend on xchecker-utils
-fn ensure_dir_all(path: &Utf8PathBuf) -> Result<(), io::Error> {
-    if !path.as_std_path().exists() {
-        fs::create_dir_all(path.as_std_path())?;
-    }
-    Ok(())
-}
-
-/// Set a thread-local override for XCHECKER_HOME during tests.
-#[cfg(any(test, feature = "test-utils"))]
-pub fn set_thread_home_for_tests(path: Utf8PathBuf) {
-    THREAD_HOME.with(|tl| *tl.borrow_mut() = Some(path));
-}
-
-/// Set up an isolated home directory for testing.
-///
-/// This avoids process-global environment changes by using thread-local state.
-#[cfg(test)]
-pub fn with_isolated_home() -> tempfile::TempDir {
-    let td = tempfile::TempDir::new().expect("Failed to create temp dir");
-    let p = Utf8PathBuf::from_path_buf(td.path().to_path_buf()).unwrap();
-    set_thread_home_for_tests(p);
-    td
-}
-
-impl XCheckerLock {
-    /// Create a new lockfile with current context
-    #[must_use]
-    pub fn new(model_full_name: String, claude_cli_version: String) -> Self {
-        Self {
-            schema_version: "1".to_string(),
-            created_at: Utc::now(),
-            model_full_name,
-            claude_cli_version,
-        }
-    }
-
-    /// Detect drift between locked values and current run context
-    /// Returns None if no drift detected, Some(LockDrift) if drift exists
-    #[must_use]
-    pub fn detect_drift(&self, current: &RunContext) -> Option<LockDrift> {
-        let mut drift = LockDrift {
-            model_full_name: None,
-            claude_cli_version: None,
-            schema_version: None,
-        };
-
-        // Check model drift
-        if self.model_full_name != current.model_full_name {
-            drift.model_full_name = Some(DriftPair {
-                locked: self.model_full_name.clone(),
-                current: current.model_full_name.clone(),
-            });
-        }
-
-        // Check Claude CLI version drift
-        if self.claude_cli_version != current.claude_cli_version {
-            drift.claude_cli_version = Some(DriftPair {
-                locked: self.claude_cli_version.clone(),
-                current: current.claude_cli_version.clone(),
-            });
-        }
-
-        // Check schema version drift
-        if self.schema_version != current.schema_version {
-            drift.schema_version = Some(DriftPair {
-                locked: self.schema_version.clone(),
-                current: current.schema_version.clone(),
-            });
-        }
-
-        // Return None if no drift detected
-        if drift.model_full_name.is_none()
-            && drift.claude_cli_version.is_none()
-            && drift.schema_version.is_none()
-        {
-            None
-        } else {
-            Some(drift)
-        }
-    }
-
-    /// Load lockfile from spec directory
-    pub fn load(spec_id: &str) -> Result<Option<Self>, io::Error> {
-        let lock_path = Self::get_lock_path(spec_id);
-
-        if !lock_path.exists() {
-            return Ok(None);
-        }
-
-        let content = fs::read_to_string(&lock_path)?;
-        let lock: Self = serde_json::from_str(&content)
-            .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?;
-
-        Ok(Some(lock))
-    }
-
-    /// Save lockfile to spec directory
-    pub fn save(&self, spec_id: &str) -> Result<(), io::Error> {
-        let lock_path = Self::get_lock_path_utf8(spec_id);
-
-        let json = serde_json::to_string_pretty(self)
-            .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?;
-
-        write_file_atomic(&lock_path, &json).map_err(io::Error::other)?;
-
-        Ok(())
-    }
-
-    /// Get the path to the lockfile for a spec ID
-    fn get_lock_path(spec_id: &str) -> PathBuf {
-        Self::get_lock_path_utf8(spec_id).into_std_path_buf()
-    }
-
-    /// Get the UTF-8 path to the lockfile for a spec ID
-    fn get_lock_path_utf8(spec_id: &str) -> Utf8PathBuf {
-        spec_root(spec_id).join("lock.json")
-    }
-}
-
-impl FlowLock {
-    /// Create a new flow lock from the current governed flow context.
-    #[must_use]
-    pub fn new(context: FlowContext) -> Self {
-        Self {
-            schema_version: "1".to_string(),
-            created_at: Utc::now(),
-            flow_version: context.flow_version,
-            stage_graph_version: context.stage_graph_version,
-            gate_set_version: context.gate_set_version,
-            prompt_pack_version: context.prompt_pack_version,
-            adapter_version: context.adapter_version,
-            provider_policy: context.provider_policy,
-            tool_context_version: context.tool_context_version,
-        }
-    }
-
-    /// Detect drift between the locked flow boundary and the current flow context.
-    #[must_use]
-    pub fn detect_drift(&self, current: &FlowContext) -> Option<FlowLockDrift> {
-        let mut drift = FlowLockDrift::default();
-
-        if self.flow_version != current.flow_version {
-            drift.flow_version = Some(DriftPair {
-                locked: self.flow_version.clone(),
-                current: current.flow_version.clone(),
-            });
-        }
-
-        if self.stage_graph_version != current.stage_graph_version {
-            drift.stage_graph_version = Some(DriftPair {
-                locked: self.stage_graph_version.clone(),
-                current: current.stage_graph_version.clone(),
-            });
-        }
-
-        if self.gate_set_version != current.gate_set_version {
-            drift.gate_set_version = Some(DriftPair {
-                locked: self.gate_set_version.clone(),
-                current: current.gate_set_version.clone(),
-            });
-        }
-
-        if self.prompt_pack_version != current.prompt_pack_version {
-            drift.prompt_pack_version = Some(DriftPair {
-                locked: self.prompt_pack_version.clone(),
-                current: current.prompt_pack_version.clone(),
-            });
-        }
-
-        if self.adapter_version != current.adapter_version {
-            drift.adapter_version = Some(DriftPair {
-                locked: self.adapter_version.clone(),
-                current: current.adapter_version.clone(),
-            });
-        }
-
-        if self.provider_policy != current.provider_policy {
-            drift.provider_policy = Some(DriftPair {
-                locked: self.provider_policy.clone(),
-                current: current.provider_policy.clone(),
-            });
-        }
-
-        if self.tool_context_version != current.tool_context_version {
-            drift.tool_context_version = Some(DriftPair {
-                locked: self.tool_context_version.clone(),
-                current: current.tool_context_version.clone(),
-            });
-        }
-
-        if drift.flow_version.is_none()
-            && drift.stage_graph_version.is_none()
-            && drift.gate_set_version.is_none()
-            && drift.prompt_pack_version.is_none()
-            && drift.adapter_version.is_none()
-            && drift.provider_policy.is_none()
-            && drift.tool_context_version.is_none()
-        {
-            None
-        } else {
-            Some(drift)
-        }
-    }
-
-    /// Load `flow.lock` from the spec directory.
-    pub fn load(spec_id: &str) -> Result<Option<Self>, io::Error> {
-        let lock_path = Self::get_lock_path(spec_id);
-
-        if !lock_path.exists() {
-            return Ok(None);
-        }
-
-        let content = fs::read_to_string(&lock_path)?;
-        let lock: Self = serde_json::from_str(&content)
-            .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?;
-
-        Ok(Some(lock))
-    }
-
-    /// Save `flow.lock` to the spec directory.
-    pub fn save(&self, spec_id: &str) -> Result<(), io::Error> {
-        let lock_path = Self::get_lock_path_utf8(spec_id);
-        let json = serde_json::to_string_pretty(self)
-            .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?;
-
-        write_file_atomic(&lock_path, &json).map_err(io::Error::other)?;
-        Ok(())
-    }
-
-    fn get_lock_path(spec_id: &str) -> PathBuf {
-        Self::get_lock_path_utf8(spec_id).into_std_path_buf()
-    }
-
-    fn get_lock_path_utf8(spec_id: &str) -> Utf8PathBuf {
-        spec_root(spec_id).join("flow.lock")
-    }
-}
-
-impl PromotionLock {
-    /// Create a new promotion lock for a promoted phase result.
-    #[must_use]
-    pub fn new(
-        spec_id: String,
-        phase: String,
-        gate_name: String,
-        gate_version: String,
-        artifacts: Vec<PromotionArtifact>,
-    ) -> Self {
-        Self {
-            schema_version: "1".to_string(),
-            promoted_at: Utc::now(),
-            spec_id,
-            phase,
-            selected_candidate_id: None,
-            gate_name,
-            gate_version,
-            approved_by: None,
-            parent_receipt_path: None,
-            parent_packet_lineage: Vec::new(),
-            artifacts,
-            warnings: Vec::new(),
-        }
-    }
-
-    /// Load the promotion lock for a specific phase.
-    pub fn load(spec_id: &str, phase: &str) -> Result<Option<Self>, io::Error> {
-        let lock_path = Self::get_lock_path(spec_id, phase);
-
-        if !lock_path.exists() {
-            return Ok(None);
-        }
-
-        let content = fs::read_to_string(&lock_path)?;
-        let lock: Self = serde_json::from_str(&content)
-            .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?;
-
-        Ok(Some(lock))
-    }
-
-    /// List all promotion locks for a spec, sorted by phase name.
-    pub fn list(spec_id: &str) -> Result<Vec<Self>, io::Error> {
-        let dir = Self::promotions_dir_utf8(spec_id);
-        if !dir.as_std_path().exists() {
-            return Ok(Vec::new());
-        }
-
-        let mut locks = Vec::new();
-        for entry in fs::read_dir(dir.as_std_path())? {
-            let entry = entry?;
-            let path = entry.path();
-            if !path.is_file() {
-                continue;
-            }
-
-            let content = fs::read_to_string(&path)?;
-            let lock: Self = serde_json::from_str(&content)
-                .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?;
-            locks.push(lock);
-        }
-
-        locks.sort_by(|a, b| a.phase.cmp(&b.phase));
-        Ok(locks)
-    }
-
-    /// Save this promotion lock under `promotions/<phase>.lock`.
-    pub fn save(&self) -> Result<(), io::Error> {
-        let lock_path = Self::get_lock_path_utf8(&self.spec_id, &self.phase);
-        let json = serde_json::to_string_pretty(self)
-            .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?;
-
-        write_file_atomic(&lock_path, &json).map_err(io::Error::other)?;
-        Ok(())
-    }
-
-    fn get_lock_path(spec_id: &str, phase: &str) -> PathBuf {
-        Self::get_lock_path_utf8(spec_id, phase).into_std_path_buf()
-    }
-
-    fn get_lock_path_utf8(spec_id: &str, phase: &str) -> Utf8PathBuf {
-        Self::promotions_dir_utf8(spec_id).join(format!("{}.lock", sanitize_lock_component(phase)))
-    }
-
-    fn promotions_dir_utf8(spec_id: &str) -> Utf8PathBuf {
-        spec_root(spec_id).join("promotions")
-    }
-}
-
-fn sanitize_lock_component(input: &str) -> String {
-    let sanitized: String = input
-        .chars()
-        .map(|ch| match ch {
-            'a'..='z' | 'A'..='Z' | '0'..='9' | '-' | '_' => ch,
-            _ => '-',
-        })
-        .collect();
-
-    if sanitized.is_empty() {
-        "unknown".to_string()
-    } else {
-        sanitized
-    }
-}
 
 /// File lock manager for spec directories
 pub struct FileLock {
@@ -1788,7 +1228,9 @@ mod tests {
         requirements.parent_receipt_path = Some("receipts/requirements.json".to_string());
         requirements.parent_packet_lineage = vec!["packet-req-1".to_string()];
         requirements.warnings = vec!["minor-style-warning".to_string()];
-        requirements.save().expect("Failed to save requirements promotion");
+        requirements
+            .save()
+            .expect("Failed to save requirements promotion");
 
         let mut design = PromotionLock::new(
             spec_id.to_string(),
@@ -1801,7 +1243,8 @@ mod tests {
             }],
         );
         design.approved_by = Some("policy-engine".to_string());
-        design.parent_packet_lineage = vec!["packet-req-1".to_string(), "packet-design-1".to_string()];
+        design.parent_packet_lineage =
+            vec!["packet-req-1".to_string(), "packet-design-1".to_string()];
         design.save().expect("Failed to save design promotion");
 
         let loaded = PromotionLock::load(spec_id, "requirements")
